feature(Settings): Further code refactoring, styling tweaks  draggale icon styling tweaks
