+++ src/containers/SettingsEditor/FormTemplates/ObjectFieldTemplate.tsx
@@ -1,18 +1,13 @@
 import clsx from 'clsx'
 import { useMemo } from 'react'
-import { Divider } from '@ynput/ayon-react-components'
 import ReactMarkdown from 'react-markdown'
 import SettingsPanel from '../SettingsPanel'
 import useCreateContext from '@hooks/useCreateContext'
 
-import { isEqual } from 'lodash'
 import { Badge, BadgeWrapper } from '@components/Badge'
 import copyToClipboard from '@helpers/copyToClipboard'
 import { $Any } from '@types'
-import { FieldTemplateProps, ObjectFieldTemplateProps } from '@rjsf/utils'
-import { CSS } from 'styled-components/dist/types'
-
-
+import { ObjectFieldTemplateProps } from '@rjsf/utils'
 
 const arrayStartsWith = (arr1: $Any, arr2: $Any) => {
   // return true, if first array starts with second array
@@ -24,11 +19,7 @@ const arrayStartsWith = (arr1: $Any, arr2: $Any) => {
   return true
 }
 
-const arrayContainsArray = (arr1: $Any, arr2: $Any) => {
-  return arr1.some((el: $Any) => isEqual(el, arr2))
-}
-
-function ObjectFieldTemplate(props: {id: string} & ObjectFieldTemplateProps) {
+function ObjectFieldTemplate(props: { id: string } & ObjectFieldTemplateProps) {
   const [contextMenu] = useCreateContext([])
   let className = 'form-object-field'
   if (props.schema.layout) className += ` layout-${props.schema.layout}`
@@ -138,14 +129,16 @@ function ObjectFieldTemplate(props: {id: string} & ObjectFieldTemplateProps) {
             )
             .map((element, index) => {
               let widget = element.content.props?.schema?.widget
-              widget = widget || ((element.content.props?.name === 'shortName') ? 'short' : null)
-              return(
-              <div key={index} className={clsx("form-object-field-item", widget && `widget-${widget}`)}>
-                {element.content}
-              </div>
+              widget = widget || (element.content.props?.name === 'shortName' ? 'short' : null)
+              return (
+                <div
+                  key={index}
+                  className={clsx('form-object-field-item', widget && `widget-${widget}`)}
+                >
+                  {element.content}
+                </div>
               )
-              }
-            )}
+            })}
         </div>
       </>
     )
@@ -224,18 +217,24 @@ function ObjectFieldTemplate(props: {id: string} & ObjectFieldTemplateProps) {
 
   if (props.idSchema.$id === 'root') {
     const projectMark = props.formContext.headerProjectName && (
-      <Badge hl="project" style={{}}>{props.formContext.headerProjectName}</Badge>
+      <Badge hl="project" style={{}}>
+        {props.formContext.headerProjectName}
+      </Badge>
     )
     const siteMark = props.formContext.headerSiteId && (
-      <Badge hl="site" style={{}}>{props.formContext.headerSiteId}</Badge>
+      <Badge hl="site" style={{}}>
+        {props.formContext.headerSiteId}
+      </Badge>
     )
 
     const envMark = props.formContext.headerVariant && (
       <Badge
-        hl={['production', 'staging'].includes(props.formContext.headerVariant)
-          ? props.formContext.headerVariant
-          : 'developer'} 
-          style={{}}   
+        hl={
+          ['production', 'staging'].includes(props.formContext.headerVariant)
+            ? props.formContext.headerVariant
+            : 'developer'
+        }
+        style={{}}
       >
         {props.formContext.headerVariant}
       </Badge>
@@ -253,9 +252,7 @@ function ObjectFieldTemplate(props: {id: string} & ObjectFieldTemplateProps) {
     )
   }
 
-  if (props.idSchema.$id === 'root') {
-    titleComponent = true ? rootTitle : stringTitle
-  }
+  titleComponent = props.idSchema.$id === 'root' ? rootTitle : stringTitle
 
   // Execute context menu
 
@@ -271,7 +268,7 @@ function ObjectFieldTemplate(props: {id: string} & ObjectFieldTemplateProps) {
       objId={objId}
       onClick={() => {
         if (props.formContext.onSetBreadcrumbs) props.formContext.onSetBreadcrumbs(path)
-      } }
+      }}
       title={titleComponent}
       description={shortDescription}
       className={`obj-override-${overrideLevel}`}
@@ -285,178 +282,4 @@ function ObjectFieldTemplate(props: {id: string} & ObjectFieldTemplateProps) {
   )
 }
 
-function FieldTemplate(props: FieldTemplateProps) {
-  const [contextMenu] = useCreateContext([])
-
-  // Do not render the field if it belongs to a different scope (studio/project/local) or if it is hidden
-
-  if (
-    props.schema.scope !== undefined &&
-    !(props.schema.scope || ['studio', 'project']).includes(props.formContext.level)
-  ) {
-    return null
-  }
-
-  const divider = useMemo(() => {
-    if (props.schema.section)
-      return <Divider>{props.schema.section !== '---' && props.schema.section}</Divider>
-    else return <></>
-  }, [props.schema.section])
-
-  // Object fields
-
-  if (props.schema.type === 'object') {
-    return (
-      <>
-        {divider}
-        {props.children}
-      </>
-    )
-  }
-
-  // Solve overrides for lists and leaves
-
-  const override = props.formContext.overrides ? props.formContext.overrides[props.id] : null
-  const path = override?.path || []
-
-  const fieldChanged = useMemo(() => {
-    return arrayContainsArray(props.formContext.changedKeys, path)
-  }, [props.formContext.changedKeys, path])
-
-  const overrideLevel = fieldChanged ? 'edit' : override?.level || 'default'
-  let labelStyle: CSS.Properties = { }
-  if (override) {
-    if (override?.inGroup) labelStyle.fontStyle = 'italic'
-  }
-
-  // Context menu
-
-  const contextMenuModel = useMemo(() => {
-    const rmPath = override?.inGroup || path
-    let model = []
-
-    if (props.formContext.onPinOverride)
-      model.push({
-        label: `Add current ${rmPath[rmPath.length - 1]} value as ${
-          props.formContext.level
-        } override`,
-        command: () => props.formContext.onPinOverride(rmPath),
-        disabled: overrideLevel === props.formContext.level,
-      })
-
-    if (props.formContext.onRemoveOverride)
-      model.push({
-        label: `Remove ${props.formContext.level} override from ${rmPath[rmPath.length - 1]}`,
-        command: () => props.formContext.onRemoveOverride(rmPath),
-        disabled: overrideLevel !== props.formContext.level,
-      })
-
-    model.push({
-      label: 'Copy',
-      command: () => copyToClipboard(JSON.stringify(props.formData, null, 2)),
-    })
-
-    model.push({
-      label: 'Paste',
-      disabled: !props.formContext.onPasteValue,
-      command: () => props.formContext.onPasteValue(path),
-    })
-
-    return model
-  }, [override, path])
-
-  const onContextMenu = (e: $Any) => {
-    e.preventDefault()
-    contextMenu(e, contextMenuModel)
-    if (props.formContext.onSetBreadcrumbs && path) props.formContext.onSetBreadcrumbs(path)
-  }
-
-  // Array fields
-
-  if (
-    props.schema.type === 'array' &&
-    // @ts-ignore
-    props.schema.items.type !== 'string' &&
-    props.schema.layout !== 'compact'
-  ) {
-    let classes = []
-
-    for (const changedPath of props.formContext.changedKeys) {
-      if (arrayStartsWith(changedPath, path)) {
-        classes.push('obj-override-edit')
-        classes.push('group-changed')
-        break
-      }
-    }
-
-    if (!classes.includes('obj-override-edit')) classes.push(`obj-override-${overrideLevel}`)
-
-    return (
-      <SettingsPanel
-        objId={props.id}
-        title={props.schema.title}
-        description={props.schema.description}
-        className={classes.join(' ')}
-        onClick={() => {
-          if (props.formContext.onSetBreadcrumbs && path) props.formContext.onSetBreadcrumbs(path)
-        } }
-        onContextMenu={onContextMenu}
-        currentId={props.formContext.currentId}
-        layout={undefined}
-        enabledToggler={undefined}
-      >
-        {props.children}
-      </SettingsPanel>
-    )
-  }
-
-  // Leaves
-
-  const widgetClass =
-    props.schema.type === 'array' && props.schema.layout === 'compact' && props.formData?.length
-      ? 'left-border'
-      : ''
-
-  // do not show error for color widgets (they are declared as strings, but
-  // contains arrays. The error is not relevant for the user)
-  //
-  // TODO: ignoring errors for now. Too many false positives
-  let className = `form-inline-field`
-  // let className = `form-inline-field ${
-  //   props.errors.props.errors && props.schema.widget !== 'color' ? 'error' : ''
-  // }`
-
-  return (
-    <>
-      {divider}
-      <div
-        className={className}
-        data-fieldid={props.id}
-        onContextMenu={onContextMenu}
-        data-tooltip={props.rawDescription}
-        data-tooltip-delay={300}
-        data-tooltip-as="markdown"
-      >
-        {props.label && (
-          <div className={`form-inline-field-label ${overrideLevel}`}>
-            <span
-              onClick={() => {
-                if (props.formContext.onSetBreadcrumbs) {
-                  if (override?.path) props.formContext.onSetBreadcrumbs(override.path)
-                }
-              }}
-              style={labelStyle}
-            >
-              {props.label}
-            </span>
-          </div>
-        )}
-        <div className={`form-inline-field-widget ${widgetClass}`}>{props.children}</div>
-      </div>
-    </>
-  )
-}
-
-
-
-export { ObjectFieldTemplate, FieldTemplate }
+export default ObjectFieldTemplate
