+++ src/containers/SettingsEditor/SettingsEditor.jsx
@@ -5,8 +5,9 @@ import { useState, useMemo, useEffect, useRef } from 'react'
 import styled from 'styled-components'
 
 import { CheckboxWidget } from './CheckboxWidget'
-import { FieldTemplate, ObjectFieldTemplate } from './Fields/fields'
-import { ArrayFieldTemplate } from './Fields/ArrayFieldTemplate'
+import FieldTemplate  from './FormTemplates/FieldTemplate'
+import ObjectFieldTemplate  from './FormTemplates/ObjectFieldTemplate'
+import ArrayFieldTemplate from './FormTemplates/ArrayFieldTemplate'
 import './SettingsEditor.sass'
 import { TextWidget } from './Widgets/TextWidget'
 import { SelectWidget } from './Widgets/SelectWidget'
